Refactoring: extract class to a separate module (OMS-223)
